+++ offering-service/src/main/java/com/moneykk/moneytown/offering/offering/query/dto/response/OfferingDetailResponse.java
@@ -5,7 +5,6 @@ import com.moneykk.moneytown.offering.offering.domain.entity.CancellationType;
 import com.moneykk.moneytown.offering.offering.domain.entity.Offering;
 import com.moneykk.moneytown.offering.offering.domain.entity.OfferingStatus;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -15,7 +14,7 @@ public record OfferingDetailResponse(
         UUID assetId,
         UUID issuerId,
         String title,
-        BigDecimal pricePerUnit,
+        Long pricePerUnit,
         Long totalQuantity,
         Long remainingQuantity,
         Long minSubscriptionQuantity,
